# perf(pnp): hoist data string