**What's the problem this PR addresses?**

https://github.com/yarnpkg/berry/pull/3977 improved the startup time of the PnP hook but we can do more.

**How did you fix it?**

Hoist the data string to the top level to avoid V8 parsing it multiple times. https://v8.dev/blog/preparser

**Benchmark results**

Gatsby benchmark:
```sh
hyperfine -w 5 "node -r ./.pnp-before.cjs -e 1" "node -r ./.pnp-after.cjs -e 1"
Benchmark 1: node -r ./.pnp-before.cjs -e 1
  Time (mean ± σ):      52.4 ms ±   1.2 ms    [User: 53.3 ms, System: 8.7 ms]
  Range (min … max):    51.1 ms …  56.4 ms    57 runs

Benchmark 2: node -r ./.pnp-after.cjs -e 1
  Time (mean ± σ):      48.5 ms ±   0.5 ms    [User: 48.0 ms, System: 9.9 ms]
  Range (min … max):    47.5 ms …  49.7 ms    61 runs

Summary
  'node -r ./.pnp-after.cjs -e 1' ran
    1.08 ± 0.03 times faster than 'node -r ./.pnp-before.cjs -e 1'
```

Next.js repo:
```sh
hyperfine -w 5 "node -r ./.pnp-before.cjs -e 1" "node -r ./.pnp-after.cjs -e 1"
Benchmark 1: node -r ./.pnp-before.cjs -e 1
  Time (mean ± σ):     134.5 ms ±   0.9 ms    [User: 144.2 ms, System: 22.8 ms]
  Range (min … max):   132.6 ms … 137.4 ms    22 runs

Benchmark 2: node -r ./.pnp-after.cjs -e 1
  Time (mean ± σ):     110.6 ms ±   0.9 ms    [User: 120.9 ms, System: 22.1 ms]
  Range (min … max):   109.4 ms … 112.5 ms    26 runs

Summary
  'node -r ./.pnp-after.cjs -e 1' ran
    1.22 ± 0.01 times faster than 'node -r ./.pnp-before.cjs -e 1'
```

**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I will check that all automated PR checks pass before the PR gets reviewed.